pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-text-outline-benchmark"
version = "0.1.0"
description = "Benchmark evaluation pipeline for ai-text-outline package"
requires-python = ">=3.9"
dependencies = [
    "psycopg2-binary>=2.9",
    "datasets>=2.14",
    "huggingface-hub>=0.19",
    "tabulate>=0.9",
    "ai-text-outline>=0.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
]
